gh-266 ChainedIterable made public
